FB-183 order events fixes

+++ backend-api/src/main/kotlin/com/rarible/flow/api/controller/OrderApiController.kt
@@ -15,5 +15,5 @@ class OrderApiController(
 ): FlowOrderControllerApi {
 
     override suspend fun getOrderByOrderId(orderId: String): ResponseEntity<FlowOrderDto> =
-        ResponseEntity.ok(service.orderById(orderId).awaitFirst())
+        ResponseEntity.ok(service.orderById(orderId.toLong()).awaitFirst())
 }
